Renamed dataformat to interfaces.ts
