+++ ts/packages/knowPro/src/index.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+export * from "./interfaces.js";
 export * from "./import.js";
-export * from "./dataFormat.js";
 export * from "./conversationIndex.js";
 export * from "./secondaryIndexes.js";
 export * from "./relatedTermsIndex.js";
