+++ ts/packages/knowPro/src/interfaces.ts
